+++ services/web/server/src/simcore_service_webserver/application.py
@@ -43,7 +43,7 @@ from .storage.plugin import setup_storage
 from .studies_dispatcher.plugin import setup_studies_dispatcher
 from .tags.plugin import setup_tags
 from .tracing import setup_app_tracing
-from .users import setup_users
+from .users.users_plugin import setup_users
 from .version_control.plugin import setup_version_control
 
 _logger = logging.getLogger(__name__)
