+++ services/web/server/src/simcore_service_webserver/exporter/formatters/formatter_v1.py
@@ -42,7 +42,7 @@ from ...director_v2.api import create_or_update_pipeline
 from ...projects.projects_api import get_project_for_user, submit_delete_project_task
 from ...projects.projects_db import APP_PROJECT_DBAPI, ProjectDBAPI
 from ...projects.projects_exceptions import ProjectsException
-from ...users_api import get_user
+from ...users.users_api import get_user
 from ...utils import now_str
 from ..exceptions import ExporterException
 from ..file_downloader import ParallelDownloader
