+++ services/web/server/src/simcore_service_webserver/groups/api.py
@@ -5,7 +5,7 @@ from aiopg.sa.result import RowProxy
 from models_library.users import GroupID, UserID
 
 from ..db import get_database_engine
-from ..users_api import get_user
+from ..users.users_api import get_user
 from . import _db
 from ._utils import AccessRightsDict
 from .exceptions import GroupsException
