+++ services/web/server/src/simcore_service_webserver/projects/_create_utils.py
@@ -21,7 +21,7 @@ from ..storage.api import (
     copy_data_folders_from_project,
     get_project_total_size_simcore_s3,
 )
-from ..users_api import get_user_name
+from ..users.users_api import get_user_name
 from . import projects_api
 from ._permalink import update_or_pop_permalink_in_project
 from ._rest_schemas import ProjectGet
