+++ services/web/server/src/simcore_service_webserver/projects/_delete_utils.py
@@ -15,8 +15,8 @@ from models_library.users import UserID
 
 from ..director_v2 import api
 from ..storage.api import delete_data_folders_of_project
-from ..users_api import UserNameDict
-from ..users_exceptions import UserNotFoundError
+from ..users.users_api import UserNameDict
+from ..users.users_exceptions import UserNotFoundError
 from .projects_db import ProjectDBAPI
 from .projects_exceptions import (
     ProjectDeleteError,
