+++ services/web/server/src/simcore_service_webserver/projects/projects_api.py
@@ -63,8 +63,8 @@ from ..socketio.messages import (
     send_messages,
 )
 from ..storage import api as storage_api
-from ..users_api import UserRole, get_user_name, get_user_role
-from ..users_exceptions import UserNotFoundError
+from ..users.users_api import UserRole, get_user_name, get_user_role
+from ..users.users_exceptions import UserNotFoundError
 from . import _delete_utils, _nodes_utils
 from .project_lock import (
     UserNameDict,
