+++ services/web/server/src/simcore_service_webserver/projects/projects_handlers.py
@@ -25,13 +25,13 @@ from servicelib.mimetype_constants import MIMETYPE_APPLICATION_JSON
 from simcore_postgres_database.models.users import UserRole
 from simcore_postgres_database.webserver_models import ProjectType
 
-from .. import users_api
 from .._meta import api_version_prefix as VTAG
 from ..director_v2.exceptions import DirectorServiceError
 from ..login.decorators import login_required
 from ..notifications import project_logs
 from ..products.plugin import Product, get_current_product
 from ..security.decorators import permission_required
+from ..users import users_api
 from . import projects_api
 from .projects_exceptions import (
     ProjectInvalidRightsError,
