+++ services/web/server/src/simcore_service_webserver/projects/projects_handlers_crud.py
@@ -37,7 +37,7 @@ from ..login.decorators import RQT_USERID_KEY, login_required
 from ..resource_manager.websocket_manager import PROJECT_ID_KEY, managed_resource
 from ..security.api import check_permission
 from ..security.decorators import permission_required
-from ..users_api import get_user_name
+from ..users.users_api import get_user_name
 from . import _create_utils, _read_utils, projects_api
 from ._permalink import update_or_pop_permalink_in_project
 from ._rest_schemas import (
