+++ services/web/server/src/simcore_service_webserver/users/users_plugin.py
@@ -12,8 +12,8 @@ from servicelib.aiohttp.rest_routing import (
     map_handlers_with_operations,
 )
 
-from . import users_handlers
-from ._constants import APP_OPENAPI_SPECS_KEY
+from .. import users_handlers
+from .._constants import APP_OPENAPI_SPECS_KEY
 
 logger = logging.getLogger(__name__)
