+++ services/web/server/tests/integration/01/test_garbage_collection.py
@@ -44,7 +44,7 @@ from simcore_service_webserver.rest import setup_rest
 from simcore_service_webserver.security.plugin import setup_security
 from simcore_service_webserver.session import setup_session
 from simcore_service_webserver.socketio.plugin import setup_socketio
-from simcore_service_webserver.users import setup_users
+from simcore_service_webserver.users.users_plugin import setup_users
 from sqlalchemy import func, select
 
 log = logging.getLogger(__name__)
