+++ services/web/server/tests/unit/with_dbs/01/studies_dispatcher/test_studies_dispatcher_projects.py
@@ -26,7 +26,7 @@ from simcore_service_webserver.studies_dispatcher._projects import (
     _create_project_with_filepicker_and_service,
     _create_project_with_service,
 )
-from simcore_service_webserver.users_api import get_user
+from simcore_service_webserver.users.users_api import get_user
 
 FAKE_FILE_VIEWS = list_fake_file_consumers()
