+++ services/web/server/tests/unit/with_dbs/01/studies_dispatcher/test_studies_dispatcher_studies_access.py
@@ -31,7 +31,7 @@ from servicelib.aiohttp.rest_responses import unwrap_envelope
 from servicelib.common_headers import UNDEFINED_DEFAULT_SIMCORE_USER_AGENT_VALUE
 from simcore_service_webserver.projects.project_models import ProjectDict
 from simcore_service_webserver.projects.projects_api import submit_delete_project_task
-from simcore_service_webserver.users_api import delete_user, get_user_role
+from simcore_service_webserver.users.users_api import delete_user, get_user_role
 
 
 async def _get_user_projects(client) -> list[ProjectDict]:
