+++ services/web/server/tests/unit/with_dbs/02/test_project_lock.py
@@ -20,7 +20,7 @@ from simcore_service_webserver.projects.project_lock import (
     is_project_locked,
     lock_project,
 )
-from simcore_service_webserver.users_api import UserNameDict
+from simcore_service_webserver.users.users_api import UserNameDict
 
 
 @pytest.fixture()
