+++ services/web/server/tests/unit/with_dbs/03/garbage_collector/test_resource_manager.py
@@ -54,9 +54,9 @@ from simcore_service_webserver.security.plugin import setup_security
 from simcore_service_webserver.session import setup_session
 from simcore_service_webserver.socketio.messages import SOCKET_IO_PROJECT_UPDATED_EVENT
 from simcore_service_webserver.socketio.plugin import setup_socketio
-from simcore_service_webserver.users import setup_users
-from simcore_service_webserver.users_api import delete_user
-from simcore_service_webserver.users_exceptions import UserNotFoundError
+from simcore_service_webserver.users.users_api import delete_user
+from simcore_service_webserver.users.users_exceptions import UserNotFoundError
+from simcore_service_webserver.users.users_plugin import setup_users
 from tenacity._asyncio import AsyncRetrying
 from tenacity.retry import retry_if_exception_type
 from tenacity.stop import stop_after_delay
