+++ services/web/server/tests/unit/with_dbs/03/test_users.py
@@ -43,16 +43,16 @@ from simcore_service_webserver.redis import (
 from simcore_service_webserver.rest import setup_rest
 from simcore_service_webserver.security.plugin import setup_security
 from simcore_service_webserver.session import setup_session
-from simcore_service_webserver.user_notifications import (
+from simcore_service_webserver.users.user_notifications import (
     MAX_NOTIFICATIONS_FOR_USER_TO_KEEP,
     MAX_NOTIFICATIONS_FOR_USER_TO_SHOW,
     NotificationCategory,
     UserNotification,
     get_notification_key,
 )
-from simcore_service_webserver.users import setup_users
+from simcore_service_webserver.users.users_models import ProfileGet
+from simcore_service_webserver.users.users_plugin import setup_users
 from simcore_service_webserver.users_handlers import _get_user_notifications
-from simcore_service_webserver.users_models import ProfileGet
 
 API_VERSION = "v0"
